[GStreamer] Remove unrealistic stalled events
Considering only webkit-network-statistics to evaluate network download stalls
isn't very realistic. On the one hand, the download may just not progress because
WebKitWebSrc may simply have decided to pause the download on purpose when it's
full of data (and that doesn't mean that the network itself is stalled).

On the other hand, when trying to detect stalls only when there's an active
download in progress, the download itself may be stalled, but there may be plenty
of buffered content to play. Therefore it would be weird that tha webpage decided
to show a spinner (because of a stalled event) when there might be a minute of
video still buffered.

To avoid this, this commit adds code to keep progressing when the download is
paused and while playback time hastn't reached the maximum time loaded.