[BCM] Force custom audiosink
The former autoaudiosink way of working, based on autoaudiosink asking
for ANY caps and then forcing the rank of brcmaudiosink wasn't very
robust and created issues with WebAudio (where a sink that understands
audio/x-raw is needed) because the Broadcom audio/x-brcm-native sink
was autoplugged.

In the end, it's better to have ifdefs that force the right audiosink
instead of playing the lottery of autoaudiosink.