Ensure that a single composition happens for each WebGL frame.
Rendering a new WebGL frame can request two compositions, one due to the layerFlush
and another one because of the new buffer pushed to the proxies. This can cause
problems to apps relying on rAF. This commit disables the composition request that
comes from pushing a new buffer from WebGL, and ensures that the only composition
request comes from the layerFlush.