[GStreamer] Add and expose AudioTrackConfiguration to JavaScript
See: https://github.com/WebPlatformForEmbedded/WPEWebKit/issues/1103

This commit is a stripped down and adapted version of
6a9975246838 which only supports AudioTrackConfiguration.

It adds a new AudioTrackConfiguration class attached to AudioTrack
and accessible from JavaScript. In this version of the patch, the
codec, sampleRate, numberOfChannels and bitrate properties are
originally stored on AudioTrackPrivate and copied to AudioTrackConfiguration
by AudioTrack when requested by updateConfigurationFromPrivate.
This happens at AudioTrack creation time, when setPrivate() is called,
or when the properties are changed in AudioTrackPrivate.

The responsibility to gather the actual configuration property values
lies in AudioTrackPrivateGStreamer, by analyzing caps and tags. In
order to do that, a notify::caps listener is installed on the pad
or stream (depending if playbin2 or playbin3 is used). TrackPrivateGStreamer
already listens to notify::tags changes, so the subclass can't install
a listener to that (it would conflict). Instead, the new virtual
TrackPrivateGStreamer::updateConfigurationFromTags() method is used
to notify AudioTrackPrivateGStreamer about changes.

AudioTrackPrivateGStreamer::updateConfigurationFromTags() takes care
of updating the bitrate. updateConfigurationFromCaps() takes care
of updating sampleRate, numberOfChannels and codec from the caps from
the pad/stream (for playbin2/3). The issue with getting the codec is
that usually the caps coming from the pad are decoded caps (audio/x-raw),
and we can't get the codec from there. The tags also show the codec,
but not with enough precision for our purpose. The solution in this
case is to navigate the pipeline back until we find an element whose
src pad has encoded caps (different from audio/x-raw). Those are what
the code calls originalCaps. The findOriginalCapsSrcPad() helper
function implements that navigation, using a PadSearchContext struct
to help the lambdas using to search the right pads.

Once we have the decoded caps, the codec_utils_caps_get_mime_codec(),
aac_caps_structure_get_audio_object_type() and
gst_codec_utils_aac_get_audio_object_type() utility functions analyze
them and produce a codec string in the format used by the HTML5 spec.

These utility functions are only available from GStreamer 1.20 on,
but we typically use GStreamer 1.16 or 1.18 with wpe-2.28. That's
the reason they've been borrowed and included in the patch on their
own.