[WTF] Distinguish SerialFunctionDispatcher that are refcounted and guarantee that the function will be run. https://bugs.webkit.org/show_bug.cgi?id=266199 rdar://119476929
Reviewed by Youenn Fablet.

A SerialFunctionDispatcher has an explicit lifetime, while most implementations
are ref-counted, their API contract do not guarantee that a dispatched function would
be run (such as a WorkerThread that can be stopped at any time).

To distinguish those, we create a RefCountedSerialFunctionDispatcher, that is both
refcounted and that has an API contract lifetime that matches its owner.

For now, only WorkQueue and RunLoop satisfy the requirement.

NativePromise is modified to take this type as the NativePromise API contract
requires that the settled method be called.

No change in observable behaviour.

* Source/WTF/wtf/FunctionDispatcher.h:
* Source/WTF/wtf/NativePromise.h:
* Source/WTF/wtf/RunLoop.h:
* Source/WTF/wtf/WorkQueue.cpp:
(WTF::WorkQueue::ref const):
(WTF::WorkQueue::deref const):
* Source/WTF/wtf/WorkQueue.h:

Canonical link: https://commits.webkit.org/271860@main